requirements.txt
# Claude MCP Init - Python Dependencies
# Version 0.11.2

# Core CLI framework
click>=8.0.0

# YAML configuration file handling
PyYAML>=6.0.1

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
types-PyYAML>=6.0.0